// happy-habitat-frontend/src/app/enums/tiempo.enum.ts
export enum MesesDelAno {
  JANUARY = 'Enero',
  FEBRUARY = 'Febrero',
  MARCH = 'Marzo',
  APRIL = 'Abril',
  MAY = 'Mayo',
  JUNE = 'Junio',
  JULY = 'Julio',
  AUGUST = 'Agosto',
  SEPTEMBER = 'Septiembre',
  OCTOBER = 'Octubre',
  NOVEMBER = 'Noviembre',
  DECEMBER = 'Diciembre'
}

export enum WeekDays {
  MONDAY = 'Lunes',
  TUESDAY = 'Martes',
  WEDNESDAY = 'Miércoles',
  THURSDAY = 'Jueves',
  FRIDAY = 'Viernes',
  SATURDAY = 'Sábado',
  SUNDAY = 'Domingo'
}

export enum HorasDelDia {
  H00_00 = '00:00',
  H00_30 = '00:30',
  H01_00 = '01:00',
  H01_30 = '01:30',
  H02_00 = '02:00',
  H02_30 = '02:30',
  H03_00 = '03:00',
  H03_30 = '03:30',
  H04_00 = '04:00',
  H04_30 = '04:30',
  H05_00 = '05:00',
  H05_30 = '05:30',
  H06_00 = '06:00',
  H06_30 = '06:30',
  H07_00 = '07:00',
  H07_30 = '07:30',
  H08_00 = '08:00',
  H08_30 = '08:30',
  H09_00 = '09:00',
  H09_30 = '09:30',
  H10_00 = '10:00',
  H10_30 = '10:30',
  H11_00 = '11:00',
  H11_30 = '11:30',
  H12_00 = '12:00',
  H12_30 = '12:30',
  H13_00 = '13:00',
  H13_30 = '13:30',
  H14_00 = '14:00',
  H14_30 = '14:30',
  H15_00 = '15:00',
  H15_30 = '15:30',
  H16_00 = '16:00',
  H16_30 = '16:30',
  H17_00 = '17:00',
  H17_30 = '17:30',
  H18_00 = '18:00',
  H18_30 = '18:30',
  H19_00 = '19:00',
  H19_30 = '19:30',
  H20_00 = '20:00',
  H20_30 = '20:30',
  H21_00 = '21:00',
  H21_30 = '21:30',
  H22_00 = '22:00',
  H22_30 = '22:30',
  H23_00 = '23:00',
  H23_30 = '23:30'
}